refactor/rework_service_launchers
wrap __main__py logic into helper classes for more flexibility
